+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **General Changes**
   - Default ratings are now set at 3 stars for unreviewed datasets and tools
+- **Admin Changes**
+  - Added a quarterly download report for admins
 - **Dataset Changes**
   - `md`, `pdf`, and image files are now displayed in browser when clicked on
     the files index